ArbeitsDialog: der Wert für die Walze wird ausschliesslich aus dem ersten Rezept hergenommen Dialoge: NewDialog: Umrechnung der Z-Achsen Werte in uint bei der Direkteingabe

+++ Dialoge.h
@@ -204,7 +204,7 @@ namespace EuMax01
     int sdlw;
     int camw;
 
-    unsigned short usWerte[6];
+    unsigned int usWerte[6];
 
     void setNewPositionValue(int pos, unsigned int value,char * suffix);
     void setNewPositionValue(int pos, unsigned int value,char * suffix,unsigned int zfaktor);
